pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spec2cov"
version = "0.1.0"
description = "Spec-to-coverage dataset pipeline for SystemVerilog/UVM repositories"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Claude Code" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "sqlalchemy>=2.0.29",
  "httpx>=0.27.0",
  "tenacity>=8.2.3",
  "datasketch>=1.6.5",
  "pyverilog>=1.3.0",
  "lxml>=5.2.1",
  "openpyxl>=3.1.2",
  "markdown-it-py>=3.0.0",
  "structlog>=24.1.0",
  "orjson>=3.10.3",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "pypdf>=5.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1",
  "pytest-cov>=5.0.0",
]

[project.scripts]
spec2cov = "spec2cov.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
